change table name discrepancies, questions
I just wrote in some questions, obviously they are intended to be reviewed and deleted, not updated to this file.  If this is an annoying or inappropriate way to do this, please let me know.  I just think it should be reviewed by an instructor -- it's beyond my capacity in this case to suggest what should or shouldn't actually be there.